Specialise voyeur manifold to stateconfigwatcher
The voyeur manifold was created to be used with the state manifold but
wasn't suitable as is because it would cause the state manifold to
restart for *any* agent config change. We only want the state manifold
to be bounced if there's a change in state serving info.

The stateconfigwatcher manifold watches the agent config and only
bounces itself when state serving info is added or removed.