cmd/juju/backups: fix issue 1546826
Fixes LP # 1546826

Remove the duplicate configuration of the logger in the backups
subcommand.

There is no test because of the way that the logging configuration
passed to the test runner interacts with logger -- in debug mode the
warning logger is not redefined, but in real use it failed.

This has been manually tested.

We should have an integration test asserting that backup and restore
work as specified.