cmd/jujud/agent/machine: Introduce stateworkers manifold
"stateworkers" is a temporary manifold that runs workers which have
not yet been converted to run directly under the dependency engine. It
obtains a *state.State from the "state" manifold and passes it to a
function which starts the workers.

As part of this change the old "state-starter" and "state" workers
have been removed from the machine agent. They have been replaced by
equivalent dependency engine manifolds.