It compiles against the proposed lxd branch.
It doesn't pass any test suites. In fact, it segfaults. But it does compile.

This rips out all the code that dealt with writing to disk in order to get
certificates and configuration working with the LXD code.
The nice part is that it is removing a bunch of code. Bad part is it doesn't work yet.