all: change bootstrap model name to "admin"
When bootstrapping, we now call the initial
model "admin". Some more commands have been
updated to stop using configstore, and more
legacy code has been removed from modelcmd
as a result.

JujuConnSuite had to be updated to name its
initial model "admin" too, to support several
commands which now assume that is what the
admin model is called. As a result, many tests
had to be updated to stop referring to the
initial model as "dummymodel".