Block major upgrade unless using compatible client
Major upgrades from 1.x to 2.x are not yet supported.

This PR backs out the commits to allow a major upgrade
in the client, and adds in an extra field in params for
SetEnvironAgentVersion that indicates a major upgrade
can be attempted.

Clients will not be updated to set that field until we
support upgrades from 1.x to 2.x