1550770: Add DefaultSeries tools for current arch
TestContainerStartedAndStopped starts a container with
the default series, but does not upload tools for non
amd64 arches.  As a result the tests fail because
tools cannot be found.

This patch uploads tools for the default series and
the current arch if the arch is not amd64 (as amd64
tools with the default series are already uploaded
in the base suite)