apiserver/common: standardise on juju/errors
common/errors.go imported both the go errors package, and our
juju/errors superset. This was so sentinel errors were created without
the usual juju/errors file/line tracking.

This is unncessary, juju/errors.New is fully compatible with errors.New;
its sentinel values can still be compared by value.

The only visible change is the error trace, rather that listing no
file/ilne for the final error value, it will list the file/line that the
value was declared at the package level. This distinction is not
important, because what we look for in the error trace is not where the
error was created, but who was the caller -- if we wanted the former,
sentinal errors are not the right tool.