provider/maas: add option to bridge all or only a specific interface
Add an option that controls whether the add-juju-bridge script (which is
run during cloud-init) should bridge all interfaces or a specific
interface only. This commit changes the default to bridging a specific
interface only and that interface is the default IPv4 route of the host
machine.

These changes are to faciliate the unblocking of CI testing on master
and these changes will eventually go away once the full multi-NIC
container solution is merged from maas-spaces2.

Fixes [LP:#1549545](https://bugs.launchpad.net/juju-core/+bug/1549545)