state: unexport UpgradeInProgressError
There are too many ways for juju's API server to tell the client that it
is in upgrade state. Some come from state itself, others come from the
API server interpreting other errors from other sources.

This CL is part of a pair of changes that reduce the number of ways the
API server can say it is in upgrade state. Specifically other packages
can only inspect the error they get from state with the supplied helper
method, they cannot inspect it directly.

This CL is less than ideal is state depends on apiserver/params for the
CodeUpgradeInProgress constant. But this source dependency existed
before hand, so it's not the end of the world.