Fix intermittent test bug #1552589.
It was possible to change the configuration before the provisioner task
was started, and thus there was never any change to observe.
Now we always observer the initial config. It means the test itself
is slightly weaker, because sometimes we never observer the *change*,
but at least the test doesn't accidentally fail.
A better fix would be to wait to change the config until you know
that the config has already been seen. However, we can't do that
via the Observer interface, because there is a race that we
could have already started the Task before we get a chance to
set the observer.