provider/maas: make bridge script idempotent
If an interface was already bridged the script would create a stuttered
iface name and an additional bridge_ports option which is invalid. For
example:

  iface eth0 inet manual

  iface test-br-eth0 inet manual

  auto test-br-test-br-eth0
  iface test-br-test-br-eth0 inet static
    address 1.2.3.4
    netmask 255.255.255.0
    gateway 4.3.2.1
    bridge_ports eth0
    bridge_ports test-br-eth0

This is fixed by only adding a bridge if the iface does not have an
existing 'bridge_ports' option.

Fixes [LP:#1553915](https://bugs.launchpad.net/juju-core/+bug/1553915)