Remove LXD LinkLocal and MachineLocal addresses.
fixes-bug#1552423
This changes the LXD client code so that calling Addresses no longer returns
local addresses. I tried just setting their Scope appropriately, but it
seems that we'll still treat them as addresses that we will try to use
for things like SSH.
We know that these addresses won't be accessible outside of the local
machine, so there is no reason for a Provider to be returning them.
We should still see them when we look at SetMachineAddresses, but that
is fine, as we handle those addresses correctly.