Angry rant time!
UpgradeInProgressError used to be defined in apiserver/admin.go and used
in apiserver/upgrading_root.go. __BUT__, the logic that turns errors
into rpc messages is inside apiserver/common/error.go. This creates a
potential import loop as apiserver already imports common, so common
cannot import apiserver to get a type.

So, to ~~solve~~ workaround this clusterfuck, we move
UpgradeInProgressError to apiserver/params, which everyone imports and
solve the import loop.

This is _SHIT_. apiserver and apiserver/common are clearly the same
package, thay cannot function without each other and have been
bifurcated for reasons that no longer make sense.

Once this PR is landed, I will merge the two packages and then we can
define, and use UpgradeInProgressError in _one_ place, and then we can
also unexport it because things should not be checking for it explicitly
on that apiserver side of the blood/brain RPC layer.