apiserver: remove dependency on resource/api
The apiserver was importing resource/api to get one constant to
construct a URL path. This made the apiserver depend on all the code
that the resource/api package depends on, and flirted with an import
loop.

Given that the line directly below the one affected by this change does
the same thing, but without a constant in another package, it seems
simpler to remove the dependency and just write the literal string in
the one place it is used, and has test coverage.