all: remove configstore
We move bootstrap config to jujuclient, and in
doing so, remove the last vestige of configstore.

Bootstrap config has been changed significantly
at the same time. We no longer store credentials
in bootstrap config. Rather, we expect the user
to maintain their credentials.yaml file, and we
store the name of the credential that was used
to bootstrap. When we request bootstrap config,
we will extract them out of credentials.yaml as
we do at bootstrap time. If credentials were
detected at bootstrap, we will detect them
again when bootstrap config is requested.

To enable this enhancement we have had to change
EnvironProvider once again. PrepareForBootstrap
has been split into two methods: BootstrapConfig
and PrepareForBootstrap. The former will take the
parameters that were previously passed to
PrepareForBootstrap, and return a complete config.
PrepareForBootstrap now just opens the complete
config and performs any required one-time
initialisation of the bootstrap environment.

BootstrapConfig is required to return the complete
config deterministically, because we do not store
the complete config on disk.