container/lxc: hard code loop device manager
lxx.NewContainerManager requires a loop device manager. This parameter
is not documented, and there is no check to ensure that it is not nil.
There is only one caller of this method, inside container/factory which
always passes a valid value. This parameter tf. is only used for testing.

This change hard codes the parameter unconditionally passed in via the
the container factory and adds a new test scoped method to alter the
loopback manager for testing.