core/description: Work around go-yaml limitation under Go 1.6
Under Go 1.6, go-yaml panics when it encounters private embedded
types. This is fine in older Go versions and is even fine on Go 1.6
with encoding/json's serialization. go-yaml needs changing and I have
some idea of the changes required but I don't have the time to make
the changes nor them merged upstream. The YAML format produced is
unaffected by this change.

For now, various private embedded types have been made public. This
can be reversed at some point in the future.

This issue is a non-obvious side-effect of
https://golang.org/doc/go1.6#reflect.