Initial work on fixing our test suite to actually have valid lifetimes for PatchValue objects.
We have a bunch of tests that thought they were patching things out and could trust them
to be correct, but they didn't realize the PatchValue lifetime was not what they thought
it was. So they were actually testing the wrong thing.
Or leaving global state accidentally dirty and the other tests then are using the wrong
data.
Our BaseSuite was guilty of this, which means lots of things could be wrong.