fixes-1558061 by skipping messages
It seems that calling SetInstanceStatus() takes about 1s for every
request. And we'll end up with 100 of them when copying an Ubuntu image.
However, that 100s is much longer than it takes to get the image on
Amazon, which means we're still reporting progress while we've moved on
to other things.

We'll want to fix this with more direct tested code, I think, but this
is a good start.