container/lxd: switch from default profile to custom network profile
Use the supplied network configuration when creating a container. This
switches away from just using the "default" LXD profile which has one
NIC and creates a custom profile based on the containers' networking
setup.

For example:

$ lxc profile show juju-machine-0-lxd-1-network
name: juju-machine-0-lxd-1-network
config: {}
description: ""
devices:
  eth0:
    mtu: "1500"
    name: eth0
    nictype: bridged
    parent: br-eth0
    type: nic
    hwaddr: 00:16:3e:83:c0:3e
  eth1:
    mtu: "1500"
    name: eth1
    nictype: bridged
    parent: br-eth0.10
    type: nic
    hwaddr: 00:16:3e:d6:3b:bf
  eth2:
    mtu: "1500"
    name: eth2
    nictype: bridged
    parent: br-eth0.11
    type: nic
    hwaddr: 00:16:3e:28:1a:dc
  eth3:
    mtu: "1500"
    name: eth3
    nictype: bridged
    parent: br-eth0.12
    type: nic
    hwaddr: 00:16:3e:98:8a:a8

A new profile will be created per container as the MAC addresses will be
different. And the relationship with eth<N> and the parent device may
also be different.

This changes realises multi-NIC container support for LXD and is
dependent on the maas-spaces-multi-nic-containers branch.