cmd/juju/commands: refactor bootstrap retry logic
Refactor bootstrap retry logic in preparation for making it a general
function of any API call.

This refactor is made a little more complex by the contract for
bootstrap that says the method will return the last error in the retry
loop -- but at least we wrap it with an explanation of what caused the
error.

Tested with various manual boostrap scenarios including terminating the
machine during bootstrap.