bootstrap: fix authorized-keys in hosted model
If --config contains authorized-keys, the system
key will not be added to the initial hosted model's
config. This is because authorized-keys is modified
in the admin model config, but that value is not
copied over to the hosted model config as it is
already present.

Fixes https://bugs.launchpad.net/juju-core/+bug/1461561