cmd/juju: add bootstrap config to show-controller
Add bootstrap config to the output of "show-controller".
This provides a more extended description of a controller,
on the client that bootstrapped it only. This includes
information about the cloud, and the credential used to
bootstrap.

Generally speaking, one should use "get-model-config"
to inspect model properties. This change is made available
only to provide details on the controller itself, and not
any models within it. It is not safe to assume that all
models within a controller have the same cloud type, as
we may in the future support heterogenous environments.

Possible for for https://bugs.launchpad.net/juju-core/+bug/1554721