cmd/...: juju errors for humans
Update the error messages that are printed out
whenever a current controller or model is not set,
and not specified. Give advice on what commands
and flags to use to rectify the situation. Give
a bit more feedback in "juju add-user" and
"juju register", and set the current model in
"juju register" if there is only one model
available.