Add direct tests of EnsureImageExists
We weren't able to directly put lxd.Client.CopyImage into an interface
because it needs the exact lxd.Client object. But if we wrap it a bit
deeper we can actually do some testing on the logic inside of
EnsureImageExists. It is a Stub (mock) sort of test which is fairly tied
to the specifics, but we really do want to test things like "what do we
do if we fail to connect to a Remote", and "what if the remote doesn't
have the Alias".

This doesn't change any functionality (that will be coming in later
patches), but does make it possible to test things.