rpc: refactor NewConn into Client and Server methods
The underlying rpc.Conn is bidirectional, but in our use case we always
send requests from the client to the server which responds.

Hide rpc.NewConn and replace it with a pair of NewClientConn and
NewServerConn methods to deliniate the method sets of both types so they
can be mocked.