1538735: Pass BootstrapSeries to provider bootstrap
--bootstrap-series is used to filter tools before
calling the provider bootstrap method, but then the
bootstrap-series is not passed into the provider when
calling bootstrap.

The provider then used either the specified
default-series, or the default default-series to choose
tools.  This led to a "no matching tools available"
error if the bootstrap series differed from the default
series used.

In this patch, I update the BootstrapParams to include
the bootstrap series and have the provider use that
to check for tools.