worker/apiconfigwatcher: Add dependency for API address changes
The apiconfigwatcher manifold monitors the agent configuration for
changes to the API addresses. For any material change the manifold
worker will restart itself.

It is intended for use as a dependency for the api-caller worker. The
apiconfigwatcher will cause the api-caller to restart whenever the
migration minion updates the API server addresses in the agent
configuration.