cmd/juju/user: add "logout" command
"juju logout" will remove the user credentials
for the specified/current controller. If there
is a password but no macaroon recorded for the
user, then the user must either acquire a macaroon
by changing the password first, or use the --force
flag.

We now disallow logging in as multiple users
for a controller simultaneously, and consequently
drop the switch-user command. You must first log
out before logging in as another user.