provider/vsphere: fix copy mutex by value error
The pattern `env = env.getSnapshot()` was broken as it would return a
shallow copy of the environ value, including its lock. However all the
fields referenced by pointers would point to the same instnace, but with
the lock no longer protecting access.

The `env = env.getSnapshot()` pattern is identical to declaring the
method on on `env`, not `*env`, but the pattern hid this incorrect
behaviour until a recent improvement to go vet spotted it.