cmd/juju/model: add "show-model" command
We add the "show-model" command which shows
summary information about the current model.
This includes the name, UUID, provider type,
owner, and users with access.

If you are an administrator or owner of the
model, you will see all users with access to
the model. Otherwise, you will see only your
own user details.