MAAS2 implementation of maasEnviron.NetworkInterfaces
Reimplement maasObjectNetworkInterfaces as maas2NetworkInterfaces.

I initially tried to share the big loop that makes InterfaceInfos by
making maasInterface &c conform to the corresponding gomaasapi
interfaces, but that required renaming most of the fields in the
structs (which leaked out into other code) and tied the v1 and v2 code
together too much.

Discarded version here:
https://github.com/babbageclunk/juju/tree/maas2-interfaces-bad-idea

This doesn't build at the moment - it relies on some fields not in the
gomaasapi yet.