Implement the releaseNodes retry for MAAS 2
On rereading I realised I'd missed a lot of the work the v1 version of
releaseNodes does. Renamed releaseNodes to releaseNodes1, and wrote
releaseNodes2 that does the same for MAAS 2 but with the new error
types.

I tried to use a type switch for the error handling, but the gomaasapi
error types don't implement Error() which seems odd.