Update help text per Doc Team recommendations
See details in https://goo.gl/MuP5SO

Also, note this commit removes a test for a lower case first letter in
the summary. Since they are all changing to upper case we can put this
back once done. Otherwise the tests will fail on merge and we won't be
able to build.