1568390 - Specify a mongo fallback for wily too
The addition of a fallback package for mongo didn't
specify a fallback for wily, so users couldn't bootstrap
a wily controller.

The test added were tested on ubuntu and centos.  This
change should land after the utils dependency is updated
to pull in https://github.com/juju/utils/pull/204

The utils change eliminates retries if a package fails
to install because it doesn't exist.