1567683: filter bridge addresses for lxd instances
The lxd provider should not return bridge addresses
from an Instance.Addresses() call.  These will be
filtered out now.

We cannot rely on FilterLXCAddresses as that filters
out addresses based on the current lxc-net config of
the controller.

I also updated the test to use a 10.0.8.* IP for eth0
to match what is currently being suggested as the
network for lxd, so that if there are filtering
problems later on, they might be caught in this test.