provider/azure: honour agent mirrors
We stopped implementing HasRegion when we moved over
to the Azure Resource Manager API, because we don't
(yet) have usable image simplestreams metadata. This
interface is also required to support agent mirrors,
though.

We introduce a new interface, HasAgentMirror, which may
be implemented instead of HasRegion; it will only be
used when looking up agent simplestreams metadata.
This is a temporary change, which we will replace with
an implementation of HasRegion for Azure when the image
metadata is available.

Fixes https://bugs.launchpad.net/juju-core/+bug/1551779