state: Don't use model UUID for GridFS namespace
In most uses of GridFS within Juju, the model UUID was being used as
the GridFS namespace. This meant that if the same resource was written
out through multiple models, it would only be accessible via the first
model (due to deduping). A fixed namespace is now used to allow
sharing of resources with identical content across models.

Fixes LP #1569054.

Note that due to the change in the underlyin gstorage locations, this
revision will break upgrades from previous 2.0 pre-releases.