Add model status
Models will now have a status document. When a model
is created, it will be "active", or "importing". When
a model progresses to Dying, its status will be set
to "destroying"; when it progresses to Dead, its status
will be set to "archived".

"juju list-models" now filters out Dead models, unless
the --archive flag is passed; and shows the current
status for models it displays.

"juju show-model" shows the lifecycle state and current
status for the model.

We don't yet rename models when they become Dead. That
will be done separately. When model migration is ready,
we should set the model status to "exporting" at the
appropriate time.

Fixes https://bugs.launchpad.net/juju-core/+bug/1534627