provider/lxd: fix off-by-1 LXD subnet calculation
The call to detectSubnet() was always adding 1 to the network address
that was discovered when parsing the output from `ip addr'. This has the
unfortunate effect of changing the subnet for the LXD bridge after it
has been recorded by Juju. This meant that further container <=>
apiserver communication would break as the original value was recorded
in the agent.conf file.

Fixes [LP:#1569361](https://bugs.launchpad.net/juju-core/+bug/1569361)