Disallow upload-tools to be used with admin model.
To be able to test this I had to add a global closure which is patched out in several tests that attempt to perform an upgrade on the admin model with an --upload-tools flag. I really didn't want to do this, but I couldn't spend any more time trying to get a non-admin model into the tests that needed it.