Move life/status out of params.Model
list-models now gets the list of models,
and then passes the tags into the ModelInfo
API call for more details. It would be ideal
to make just one API call, but there is no
time to do that right now.

The YAML/JSON format of list-models now
contains all of the information included
in "show-model", but is not presented in
a map format. The reason for that is that
we may be listing models for multiple users,
which may each have models with the same
name.