worker/uniter/storage: initialise context
The existing logic for initialising storage
attachment context was a bit broken, because
it was a mish-mash of the old pre-resolver
code, and new resolver logic.

Now we ensure that:
 - when the uniter starts up, it will initialise
   the context for any attachments it previously
   had issued storage-attached for.
 - we update context immediately before running
   a hook, and only for the storage related to
   that hook.

Much of the old code is removed, because after
initialisation we should only refer to the
resolver's remotestate.Snapshot for guidance
on what to do next, and how to update local and
remote state.

Fixes https://bugs.launchpad.net/juju-core/+bug/1569490