lxd: get rid of fallback to lxcbr0
The fallback was buggy: lxdbr0 no longer starts automatically any more, so
we can't assume that statting the file exists.

This code is all mostly useless anyway as it was mostly for the transition
from lxcbr0 to lxdbr0. Let's just get rid of any concept of falling back
and just do what we did before.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>