worker/peergrouper: remove serverInfoWatcher
The peergrouper now manages the two controller watchers more directly,
removing the need for serverInfoWatcher and removing one of the uses
of the crazy notifyCh mechanism (going soon). The updateMachines logic
is now a method on the peergrouper worker as it always should have
been. The watchers are now managed by the catacomb, which removes some
concerns from the peergrouper.