Fix InstanceConfig.SetTools usage
InstanceConfig.SetTools was being strict about requiring tools
to have URLs, which is not the case when using --upload-tools.
It may be worth revisiting this later, but for now we revert
to not checking the URLs.

Also, fix the Azure provider to filter tools by the chosen
architecture. It was working by luck previously because the
first item in the tools list would always be amd64. If we don't
filter, then SetTools rightly throws an error because there are
multiple tools with differing architectures.

Fixes https://bugs.launchpad.net/juju-core/+bug/1571832