cloudconfig: honour full tools list
In order to properly support bringing up machines in a model
with a private network separate from the controller's, we
must attempt to download from all of the tools URLs we have
available. This means we will not only attempting each controller,
but each controller by all of their addresses.

Windows was still only attempting the first URL. This has been
fixed, so that we now attempt all of the URLs and then retry
the whole lot with a delay.

While making the change, it was found that tools catalogues
were not properly model-aware. The collection has been made
model-specific, and we have changed the tools storage to combine
the model and controller stores. This means that if you have
tools in the controller model, you can see them from the hosted
models. Any tools uploaded to a hosted model will take precedence,
and do not affect any other model.

Fixes https://bugs.launchpad.net/juju-core/+bug/1571832