1564503 - Get arm64 unit tests passing
The juju/juju/provider/openstack package is the last
failing package for the arm64 unit tests.  I added in
image data for arm64 to enable the tests to pass.

There is a test which verifies that specifying a valid
arch for which there are no images fails properly.  We
are no longer testing i386, so that is now the arch
that has no tools.