worker/provisioner: filter container tools to host
In case there are multiple agent architectures available for
starting a container, ensure we filter to the host architecture.
We were doing this for LXC, but not for LXD or KVM. We would
previously have been luck on amd64, since it's always the first
in the list.

Fixes https://bugs.launchpad.net/juju-core/+bug/1572781