container/lxd: only write bridge config once
Every time [container/lxd] Initialise() was run (reboot, upgrade, etc)
the bridge configuration would be re-written with the next available
subnet. We now check to see if /etc/default/lxd-bridge has an IPv4
address set by inspecting the configuration entry `LXD_IPV4_ADDR' for an
IP address. If it is set we consider the configuration to be complete.
If it is not set (first-time install) then a new subnet is allocated.

Fixes [LP:#1569361](https://bugs.launchpad.net/juju-core/+bug/1569361)