utils/stringforwarder: remove time sensitivity
The StringForwarder was sensitive to goroutine
scheduling timing, which makes testing it unreliable.

This diff makes the StringForwarder immune to
scheduler timings by changing (clarifying?) the
contract of Forward slightly: it will not be blocked
by the callback, but it will block until the message
is forwarded to the loop, unless there is another
message currently being handled, or if the forwarder
is stopped.

The added test fails reliably for me without the
accompanying changes when run under the race detector.
No races are detected (there aren't any data races),
but the race detector causes enough overhead to
manifest the issue.

Fixes https://bugs.launchpad.net/juju-core/+bug/1570883