cmd/pprof: inject the socket path into Start
Rather than using os.TempDir() from within the
code, inject the socket path into the Start
function and leave it to the caller to decide
where to put it.

Without doing this, there is a small chance
that two tests in the same package will collide.
The tests are run sequentially, but there is a
goroutine started by the Start method that may
remove the socket after the stop function is
called, and after the proceeding test has
started listening on the socket.

Fixes https://bugs.launchpad.net/juju-core/+bug/1553292