Ignore tests for juju-1 existence on windows
AFAICT we added some checks to alert folks on xenial who might have
received an updated juju when performing a dist-upgrade. So if there is
a juju 1.x environment it warns the first time the new juju 2.x command
is run and notifies how to continue working with 1.x. If

It doesn't warn if there's 2x data, and it doesn't warn if there is
neither 1.x or 2.x data.

All checks are predicated on a command juju-1 on the path, which will
only exist on ubuntu (I think). Since this won't work on Windows which
will look for juju.exe, we skip these tests so they don't block
releases.

There may be a more correct solution for windows, but I don't have
windows to build, install, or test on.

Refs https://bugs.launchpad.net/juju-core/+bug/1564622
Refs https://bugs.launchpad.net/juju-core/+bug/1579059