Skip TestFirstRun2xFrom1xOnUbuntu on Windows
This test can't work on Windows because the Juju config directories
are the same for 1.x and 2.x on Windows and the code being tested
requires that they be different. Given that the section of code being
tested doesn't need to run under Windows the test can be skipped
there.

Fixes LP #1580314.