state: re-enable tests under race detector
Fixes LP 1519095

The good news, there are no data races in state currently.
The bad news, the tests take about 1.5x the time to run. On my machine
it is the difference between 15 minutes and 25 minutes to run the -race
tests for this package. Those times will be higher if other tests are
run in parallel (which is the default mode when testing multiple
packages).