apiserver: fix data race on tls config
Fixes LP 1518806
Fixes LP 1519097

Pass a _copy_ of the current TLS config to the TLS server for the
accepted connection. This fixes a race where updateCertificate may
mutate its config, a reference to which has been passed to _all_ the
active TLS connections.

This bug is at the bottom of a chain of other data race blockers so we
can remove a bunch of test skips under the race detector.