provider/azure: don't log unknown sizes as errors
It's not an error if we don't know a VM size. The
sizes that Azure supports grow over time, and the
Juju code will need to be updated eventually to
support them completely. In the mean time we just
say the unknown ones are more expensive than the
ones we know about, so they're not chosen without
explicit request.