Add cloud/region to controller details
Add the name of the cloud and region (if applicable)
to the controller details. This is included in the
output of list-controllers and show-controller, and
is stored in controllers.yaml.

In a follow-up, we will store cloud and credential
information in the controller (i.e. in mongo), which
will enable us to display the same information for
"registered" controllers. For now, registered
controllers lack that information.