state: fix data race in all watcher
Fixes LP 1586244

Fix a data race in the tests where they assumed they had ownership of
various multiwatcher.EntityInfo values. This was not true, and was a
violation of the contract of the multiwatcher's cache Get method.

This is not a perfect fix -- the ideal would be for the cache to not
require that values returned from Get be unmolested, but that isn't
going to happen any time soon.

Ideally we'd drop this "mash some status fields and use deep equals" in
the tests, but this assumption is woven through the test heavily.

So the best alternative is to try to make overwriting of the status
fields non destructive, which _should_ work for the moment as StatusInfo
is a value type, not a reference type -- famous last words and ones I
will surely regret at some point in the near future.

Also, include some driveby fixes to the all watcher to use the presence
value from the map, just assuming nil means "not found"