Inject iterator to raise ErrCursor directly
Start of a fix for https://bugs.launchpad.net/juju-core/+bug/1573286

Using emptycapped with an open iterator no longer raises ErrCursor from
the iterator - it gives a more generic "collection truncated" instead.

Rather than making the OplogTailer handle this error that only comes up
in testing, change the tailer to accept an iterator factory in testing
so that we can have the cursor behave the way we need.

WIP - pushing to show Menno what I've got so far.