apiserver: Open up HTTP endpoints to controller machines
There are some workers in controller machines which act on behalf on a
specific hosted model. Controller machines have been able to log into
APIs for hosted models for some time, but not Juju's HTTP style API
endpoints.

This change applies the same special casing to the authentication for
HTTP endpoints.