apiserver: fix increment of waitgroup count after wait
Fixes LP 1588143

This changes fixes a bug whereby the count of requests in fly could be
incremented after a call to wg.Wait() is issued. The contract for
wg.Wait says that once this function has been called, only calls to
wg.Done, that is deincrement, may be used.

The shutdown logic for the apiserver is now:

1. wait for srv.tomb.Dying to return.
2. close srv.listner.
3. wait for srv.wg.Wait to reach zero, implying the mongopinger has
   exited and all in fly requests have completed.

Additionally, when a request is received in the short period between the
kill of srv.tomb and the close of srv.listener, return a proper 503
response rather than just leaving the client hanging with no response.